feov(ocr): render a scan at its own resolution, and two corpus pages for what nothing covered (#1107)
* feov(ocr): a scan is rendered at its own resolution, unrounded and per axis

NativeDPIs rounded PDFium's fractional resolution to a whole number, so a scan
whose true resolution is not an integer -- which is nearly all of them -- was
RESAMPLED by us on the way in. The corpus's NBS page is 2612 px over 537 pt,
which is 350.207 DPI; we asked for 350 and got 2611x3466 from a 2612x3468
bilevel source. On a 1-bit scan every stroke edge is then interpolated, which
fills exactly the thin white counter that separates a 5 from a 9.

Measured on one cell of that page, which the scan shows as 5.4:

    resampled 2611x3466   reads 9.4   confidence 90.0
    1:1       2612x3468   reads 3.4   confidence 54.2

Both wrong -- the glyph has a broken top-bar junction and really is damaged --
but the resample also manufactured the CONFIDENCE, putting a wrong value inside
the distribution of correct ones (93-96) where no threshold could find it. I
had concluded from the resampled numbers that this error class carries no
signal. That conclusion was drawn on data this defect produced.

Two axes, not one: a scan's horizontal and vertical resolutions differ by a
hair on half the corpus, and using the horizontal for both squeezed the other
-- 3 px of vertical scaling on nbs602-form that nothing else was touching.

THE COST IS REAL AND IS NOT A WIN ACROSS THE BOARD. Over the eight corpus
pages the reading is 135 characters SHORTER: five pages lose text, one gains,
two are unchanged, and two that passed their expect block now fail it. The
reading is coherent -- tesseract's models are trained on antialiased text, and
our accidental 0.06% downscale was adding a little of it, helping most glyphs
while destroying the ones whose counters are a pixel wide. Neither position is
free; this one does not alter the evidence.

Also fixed, because the change exposed it: `lens render-page` rendered at a
fixed 300 and compared render hashes against a reading made at the page's own
resolution, so matches_reading was false for every scanned page -- a check
reporting "these differ" about a document nobody had touched.

Rule-Class: invariant-at-wrong-level
Sibling-Sweep: every carrier of a page resolution now holds the measured value
rather than a rounding -- PageRender gains WidthPx/HeightPx (the raster that
exists, not the one the caller meant), pageReceipt and PageReading carry the
exact DPI, and the fetch and ocr summaries print it to two decimals.


* feov(ocr): two corpus pages for the formats and resolutions nothing covered

The corpus was eight pages from three publishers in two encodings: CCITT G4
fax at 300, and MRC (JBIG2 text layer over a JPX background) at 350 and 501.
Three things it could not see:

  JPEG. Nothing exercised DCTDecode, the encoding a greyscale or colour scan
  usually arrives in, and nothing was 24-bit colour at all.

  The upsample path. RenderDPIFor floors a low-resolution scan at 300, and no
  page was below 300, so the floor had no specimen and the branch was reasoned
  about rather than read.

  Column layout and non-ASCII text. Every page was one or two columns of
  English; nothing had accents, and nothing had three columns.

usgs-bibliography-jpeg (US Geological Survey Bulletin 7, 1889) is a 400-DPI
24-bit colour JPEG of 1880s letterpress: a bibliography with French and German
titles, accented characters, hanging indents, and cursive marginalia running
across the foot of the page in a hand the reader has no business getting right.

gpo-record-threecolumn (Congressional Record, 1994) is a 150-DPI JPEG rendered
UP to the 300 floor: three columns of dense text with small-caps names. It is
the only page here whose golden reads "rendered at 300.00 DPI (native 150.11)".

Both read correctly today, including the accents and the three columns, so
both are green rows in the meter rather than defect specimens. That is the
point: a corpus of only the pages that beat the reader cannot show a change
BREAKING something that worked.

Still uncovered, stated rather than fabricated: no page is above the 600 cap,
so the downsample branch has no specimen.


* plans: the dashed-rule count is 10 of 12, re-measured on the publisher's bytes

The 8 came from nbs602-stripped.pdf, a re-render of the document that differs
from the original in 9.07% of its pixels. Re-run against a file that renders
pixel-identically to the publisher's, the repaired path finds 10 of the 12.


---------

+++ plugins/frank-exchange-of-views/tools/internal/cli/fetchsummary.go
@@ -77,10 +77,10 @@ type fetchSummary struct {
 	// TextCellPages counts those table pages rebuilt from the rules' own cells — a ruled table
 	// of TEXT, whose rows would otherwise arrive column by column with the row binding lost.
 	// A table page in neither count fell back to plain text, and the reading record says why.
-	TextCellPages int    `json:"text_cell_pages,omitempty"`
-	Engine        string `json:"engine,omitempty"`
-	DPILow        int    `json:"dpi_low,omitempty"`
-	DPIHigh       int    `json:"dpi_high,omitempty"`
+	TextCellPages int     `json:"text_cell_pages,omitempty"`
+	Engine        string  `json:"engine,omitempty"`
+	DPILow        float64 `json:"dpi_low,omitempty"`
+	DPIHigh       float64 `json:"dpi_high,omitempty"`
 }
 
 // applyReading folds the engine's reading of the page images into the summary.
